AMP-860: - update tool-conf.xml - update all mgm tools xml and py files with refactored file path etc - any related changes in python files
