+++ config/tool_conf.xml
@@ -8,54 +8,59 @@
     <tool file="cloud/send.xml" />
   </section>
   
+<!--  AMP MGM Sections -->  
+  
   <section id="media" name="Media Pre-Processing">
     <tool file="amp_media/extract_audio.xml"/>
     <tool file="amp_media/remove_trailing_silence.xml"/>
   </section>
   
-  <section id="silence" name="Silence Tools">
+  <section id="segmentation" name="Segmentation">
+    <tool file="amp_segment/ina_speech_segmenter.xml"/>
+    <tool file="amp_segment/remove_silence_music.xml"/>
+    <tool file="amp_segment/remove_silence_speech.xml"/>
+    <tool file="amp_segment/adjust_transcript_timestamps.xml"/>
+    <tool file="amp_segment/adjust_diarization_timestamps.xml"/>
   </section>
   
-  <section id="Transcribe" name="Speech to Text">
-    <tool file="aws/transcribe.xml"/>
-    <tool file="kaldi/kaldi_local.xml"/>
+  <section id="speech_to_text" name="Speech to Text">
+    <tool file="amp_stt/aws_transcribe.xml"/>
+    <tool file="amp_stt/kaldi.xml"/>
+    <tool file="amp_stt/vtt_generator.xml"/>
   </section>
   
-  <section id="Segmentation" name="Segmentation">
-    <tool file="ina_speech_segmenter/ina_speech_segmenter.xml"/>
-    <tool file="ina_speech_segmenter/remove_silence_music.xml"/>
-    <tool file="ina_speech_segmenter/remove_silence_speech.xml"/>
-    <tool file="amp_json_schema/adjust_timestamps.xml"/>
-    <tool file="amp_json_schema/adjust_diarization_timestamps.xml"/>
-  </section>
-  
-  <section id="nlp" name="Natural Language Processing">
-    <tool file="spacy/spacy.xml"/>
-    <tool file="aws/aws_comprehend.xml"/>
+  <section id="named_entity_recognition" name="Named Entity Recoginition">
+    <tool file="amp_ner/spacy.xml"/>
+    <tool file="amp_ner/aws_comprehend.xml"/>
   </section>
 
-  <section id="shot_detection" name="Video Shot Detection">
-    <tool file="video_tools/pyscenedetect_shot_detection.xml"/>
-  </section>
-  <section id="contact_sheet_tools" name="Contact Sheets">
-    <tool file="video_tools/contact_sheets.xml"/>
-  </section>
-
-  <section id="video_ocr" name="Video OCR">
-    <tool file="video_tools/tesseract.xml"/>
-    <tool file="video_tools/azure_video_indexer.xml"/>
-  </section>
-
-  <section id="VTTgenerator" name="JSON to VTT generator">
-    <tool file="vtt_generator/jsonToVtt_adapter.xml"/>
+  <section id="music_ocr" name="Music Optical Charater Recoginition">
   </section>
   
-  <section id="hmgm" name="Human MGM">
-    <tool file="hmgm/hmgm_sample.xml"/>
-    <tool file="hmgm/hmgm_transcript.xml"/>
-    <tool file="hmgm/hmgm_ner.xml"/>
-    <tool file="hmgm/hmgm_segmentation.xml"/>
+  <section id="video_ocr" name="Video Optical Charater Recoginition">
+    <tool file="amp_vocr/tesseract.xml"/>
+    <tool file="amp_vocr/azure_video_ocr.xml"/>
   </section>
+  
+  <section id="shot_detection" name="Shot Detection">
+    <tool file="amp_shot/pyscenedetect_shot_detection.xml"/>
+  </section>
+  
+  <section id="facial_recognition" name="Facial Recognition">
+  </section>
+  
+  <section id="contact_sheet" name="Contact Sheet">
+    <tool file="amp_contact/contact_sheets.xml"/>
+  </section>
+
+  <section id="hmgm_editor" name="Human MGM Editor">
+    <tool file="amp_hmgm/hmgm_sample.xml"/>
+    <tool file="amp_hmgm/hmgm_transcript.xml"/>
+    <tool file="amp_hmgm/hmgm_ner.xml"/>
+    <tool file="amp_hmgm/hmgm_segmentation.xml"/>
+  </section>
+  
+<!--  END AMP MGM Sections -->    
   
   <section id="collection_operations" name="Collection Operations">
     <tool file="${model_tools_path}/unzip_collection.xml" labels="new" />
