+++ tools/amp_hmgm/hmgm_sample.xml
@@ -3,7 +3,9 @@
   <requirements>
 	<requirement type="package" version="3.8">python</requirement>
   </requirements>
-  <command detect_errors="exit_code">'$__tool_directory__/hmgm_sample_wrapper.py' '$output'</command>
+  <command detect_errors="exit_code">
+  	'$__tool_directory__/hmgm_sample.py' '$output'
+  </command>
   <inputs>
     <param name="input" type="data" format="json" label="Input JSON file" help="A JSON file"/>
     <param name="context_json" type="text" label="Human MGM Task Context" help="Context information needed for creating Human MGM tasks"/>
