+++ tools/amp_hmgm/hmgm_segmentation.xml
@@ -5,9 +5,9 @@
 	<requirement type="package" version="2.0">jira</requirement>
   </requirements>
   <command detect_errors="exit_code">
-	<!-- '$__tool_directory__/../amp_json_schema/segment_to_structmeta.py' '$__root_dir__' '$amp_segments_uncorrected' '$structmeta_uncorrected'; -->
+	<!-- '$__tool_directory__/segment_to_structmeta.py' '$__root_dir__' '$amp_segments_uncorrected' '$structmeta_uncorrected'; -->
 	'$__tool_directory__/hmgm_main.py' 'Segmentation' '$__root_dir__' '$structmeta_uncorrected' '$structmeta_corrected' '$task_info' '$context_json';
-	<!-- '$__tool_directory__/../amp_json_schema/structmeta_to_segment.py' '$__root_dir__' '$structmeta_corrected' '$amp_segments_uncorrected' '$amp_segments_corrected' -->
+	<!-- '$__tool_directory__/structmeta_to_segment.py' '$__root_dir__' '$structmeta_corrected' '$amp_segments_uncorrected' '$amp_segments_corrected' -->
   </command>
   <inputs>
     <param name="amp_segments_uncorrected" type="data" format="segment" label="AMP Segments Uncorrected" help="Original AMP segment file to be corrected "/>
