+++ tools/amp_ner/aws_comprehend.py
@@ -14,10 +14,9 @@ import tarfile
 from datetime import datetime
 import boto3
 
-sys.path.insert(0, os.path.abspath('../../../../../tools/amp_json_schema'))
-
-from entity_extraction_schema import EntityExtraction, EntityExtractionMedia, EntityExtractionEntity
-from speech_to_text_schema import SpeechToText, SpeechToTextMedia, SpeechToTextResult, SpeechToTextScore, SpeechToTextWord
+sys.path.insert(0, os.path.abspath('../../../../../tools/amp_schema'))
+from entity_extraction import EntityExtraction, EntityExtractionMedia, EntityExtractionEntity
+from speech_to_text import SpeechToText, SpeechToTextMedia, SpeechToTextResult, SpeechToTextScore, SpeechToTextWord
 
 def main():
     (input_file, json_file, bucketName, dataAccessRoleArn) = sys.argv[1:5]
