+++ tools/amp_ner/spacy.xml
@@ -5,7 +5,7 @@
     <requirement type="package" version="8.30">coreutils</requirement>
   </requirements>
   <command detect_errors="exit_code">
-    '$__tool_directory__/spacy_wrapper.py' '$amp_transcript' '$amp_entities' '$ignore_categories'
+    '$__tool_directory__/spacy.py' '$amp_transcript' '$amp_entities' '$ignore_categories'
   </command>
   <inputs>
     <param name="amp_transcript" type="data" format="transcript" label="AMP Transcript" help="AMP Transcript file from speech-to-text with timestamp adjusted"/>
