+++ tools/amp_stt/aws_transcribe.xml
@@ -6,8 +6,8 @@
 	<requirement type="package" version="1.5">jq</requirement>
   </requirements>
   <command detect_errors="exit_code">
-	'$__tool_directory__/transcribe.sh' '$__root_dir__/../galaxy_tmp/aws/transcribe' '$input_audio' '$aws_transcript' '$audio_format' '$s3_bucket' '$s3_directory';
-	'$__tool_directory__/../amp_json_schema/aws_transcribe_to_schema.py' '$input_audio' '$aws_transcript' '$amp_transcript' '$amp_diarization'
+	'$__tool_directory__/aws_transcribe.sh' '$__root_dir__/../galaxy_tmp/aws/transcribe' '$input_audio' '$aws_transcript' '$audio_format' '$s3_bucket' '$s3_directory';
+	'$__tool_directory__/aws_transcript_to_amp_transcript.py' '$input_audio' '$aws_transcript' '$amp_transcript' '$amp_diarization'
   </command>
   <inputs>
   	<!-- TODO
