+++ tools/amp_stt/kaldi.xml
@@ -5,8 +5,8 @@
     <requirement type="package" version="8.30">coreutils</requirement>
   </requirements>
   <command detect_errors="exit_code">
-  	'$__tool_directory__/kaldi_local.py' '$input_audio' '$kaldi_transcript_json' '$kaldi_transcript_text';
-    '$__tool_directory__/../amp_json_schema/kaldi_local_to_schema.py' '$input_audio' '$kaldi_transcript_json' '$kaldi_transcript_text' '$amp_transcript'
+  	'$__tool_directory__/kaldi.py' '$input_audio' '$kaldi_transcript_json' '$kaldi_transcript_text';
+    '$__tool_directory__/kaldi_transcript_to_amp_transcript.py' '$input_audio' '$kaldi_transcript_json' '$kaldi_transcript_text' '$amp_transcript'
   </command>
   <inputs>
     <!-- should be wav data type -->
