+++ tools/amp_vocr/azure_video_ocr.xml
@@ -5,7 +5,7 @@
 	<requirement type="package" version="2.12">azure-cli</requirement>
   </requirements>
   <command detect_errors="exit_code">
-  	'$__tool_directory__/azure_video_indexer_wrapper.py' '$input_video' '$location' '$__root_dir__' '$azure_ocr' '$azure_ocr_long' '$amp_vocr'
+  	'$__tool_directory__/azure_video_ocr.py' '$input_video' '$location' '$__root_dir__' '$azure_ocr' '$azure_ocr_long' '$amp_vocr'
   </command>
   <inputs>
 	<param name="input_video" type="data" format="video" label="Input Video" help="An input video file for OCR"/>
